panel: Ensure we dynamically update in response to UI scaling changes
The user, or the system, may decide to enforce a new scaling factor on the
UI at any given time. This may be in response to switching to a 4K output
or similar.

As such, when we detect a hidpi change we should re-assert our struts and
position. Eventually this will be handled by the PanelManager ensuring that
all panels are reset for their parent monitors and scale factor.

This will be the solution for solus-project/budgie-desktop#679

Signed-off-by: Ikey Doherty <ikey@solus-project.com>